Fix more bad tests
