+++ src/Symfony/Bridge/Monolog/Tests/Processor/DebugProcessorTest.php
@@ -66,20 +66,12 @@ class DebugProcessorTest extends TestCase
     /**
      * @group legacy
      * @expectedDeprecation The "Symfony\Bridge\Monolog\Processor\DebugProcessor::getLogs()" method will have a new "Request $request = null" argument in version 5.0, not defining it is deprecated since Symfony 4.2.
+     * @expectedDeprecation The "Symfony\Bridge\Monolog\Processor\DebugProcessor::countErrors()" method will have a new "Request $request = null" argument in version 5.0, not defining it is deprecated since Symfony 4.2.
      */
-    public function testInheritedClassCallGetLogsWithoutArgument()
+    public function testInheritedClassWithoutArgument()
     {
         $debugProcessorChild = new ClassThatInheritDebugProcessor();
         $debugProcessorChild->getLogs();
-    }
-
-    /**
-     * @group legacy
-     * @expectedDeprecation The "Symfony\Bridge\Monolog\Processor\DebugProcessor::countErrors()" method will have a new "Request $request = null" argument in version 5.0, not defining it is deprecated since Symfony 4.2.
-     */
-    public function testInheritedClassCallCountErrorsWithoutArgument()
-    {
-        $debugProcessorChild = new ClassThatInheritDebugProcessor();
         $debugProcessorChild->countErrors();
     }
 
@@ -94,18 +86,5 @@ class DebugProcessorTest extends TestCase
             'datetime' => new \DateTime(),
             'extra' => [],
         ];
-    }
-}
-
-class ClassThatInheritDebugProcessor extends DebugProcessor
-{
-    public function getLogs()
-    {
-        parent::getLogs();
-    }
-
-    public function countErrors()
-    {
-        parent::countErrors();
     }
 }
