+++ src/Symfony/Component/DomCrawler/Tests/AbstractCrawlerTest.php
@@ -1238,11 +1238,3 @@ HTML;
         return $domxpath->query('//div');
     }
 }
-
-class ClassThatInheritCrawler extends Crawler
-{
-    public function children()
-    {
-        parent::children();
-    }
-}
