+++ src/Symfony/Component/Form/Tests/TypeExtensionWithoutExtendedTypes.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Form\Tests;
+
+use Symfony\Component\Form\AbstractTypeExtension;
+
+class TypeExtensionWithoutExtendedTypes extends AbstractTypeExtension
+{
+}
